Changed and fixed how drag selecting works and added a unit manager
