+++ Assets/Scripts/Player/UnitManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: bafd1872913f69e49997611acd3bbbf7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
